clippy: get clippy to run cleanly
The majority of these are "remove &s and refs" which I think are
uncontroversial. I also removed "x: x" field assignments, which I
believe Steven didn't like in the past, but that ship has sailed.

Other controversial changes: replaced an .expect(&format!) call
with .unwrap_or_else(|_| panic!) which I guess is ok, but more
verbose.

I whitelisted the "use format!("{variable}")" lint because it's
ridiculous. This syntax doesn't work in format-like macros like
assert! or assert_eq!, and doesn't work if the 'variable' is a
field access or has a & or anything at all. So if you were to
try to make this ugly and unnecessary change, the result would
be an inconsistent mess.