clippy: set MSRV to 1.63.0
This crate compiles with 1.48.0, but depends on rust-bitcoin whose `bitcoin-io`
depenedency requires 1.63.0. My local CI tries to compile `io` even though it
isn't used.

We will need to update the MSRV even further to compile with Simplicity. But
for now get it into a state where we can compile with local CI.