run `cargo fmt` across the crate
I would like to keep imports sorted, etc etc. Do a one-time format.
It appears that `cargo fmt` and `cargo +nightly fmt` have the same
result, but it also seems like `cargo fmt` isn't trying to read
rustfmt.toml. I'm not sure what's up.

Some of the changes, in particular squishing stuff onto single lines,
are clearly for the worse and I don't know or care enough about rustfmt
to prevent them from happening. I think they're worth living with to
have the rest of the codebase formatted.

Anyway if it's too much of a pain to keep formatted we don't need to
do it on every commit.