cli tests: add regression tests for the entire CLI intereface
Let Claude 4 write a couple of these. It did a pretty bad job. Was
fairly expensive to round-trip through aider; it would make up addresses
in the wrong format; I told it to try all the flags but it didn't try -v;
it didn't use the short forms.

So the vast majority of this was a manual commit. Took me several hours.

Added a bunch of FIXMEs for things that look or feel wrong. We can revisit
them in later PRs.