hal_simplicity: attempt to decode commitnodes as hex if base64 fails
Unsure if we should move this parsing logic upstream or what to do with it.
Basically, when reading Simplicity code off of blockstream.info, it's pretty
hard to decode because the code is hex-encoded and not labeled at all and
just thrown in a big list of witness stack items.

For example:

https://blockstream.info/liquidtestnet/nojs/tx/e54d31ce544b65a3768d7dc44a9caf1142eb1ce9bb46707f5a83cb1ccf9b77f9?expand

which was created using the web ide tutorial. The program is the element
starting with e4fba05... and the witness is the element above it. The
easiest way to figure this out, for now, is to just attempt to parse all
the hex blobs as Simplicity. This commit makes that possible. (Then once
you have the hex, you can run

    hal-simplicity simplicity simplicity info <hex>

to get the base64, then re-run with other blobs

    hal-simplicity simplicity simplicity info <base64> <witness hex>

to figure out that the witness is the hex blob starting with 9bef...).