simplicity: add `sighash` command
This command takes a transaction environment (a raw transaction, input
index, details about all its input UTXOs and information about the current
input) and produces a sighash.

It also optionally takes a secret key, with which it will produce a signature
on the current sighash, and/or a public key and signature, with which it can
validate a signature. If both a public and secret key are provided, they
will be checked for consistency with each other.