simplicity: switch 'unspendable key' to the one specified in BIP 341
Our current "unspendable key" comes from the Simplicity web IDE (from
where it was copied to several other places, oops). However, this key
seems to come from an old rust-simplicity unit test where it wasn't
described as being "unspendable" and which we suspect that a private
key was once known.