add 'simplicity pset run' command which outputs jet calls and inputs and outputs
This is a bit of a hacky and incomplete implementation but it's already very
useful so I think we should ship it. It runs through a program and outputs
information about each jet that it hits. It special-cases the equality jets
and also splits up the input so you can see which values it's comparing.

It is not encoding input/output data correctly if the data length isn't a
multiple of 8 (or 64 actually); it just takes the raw words from the bit
machine and outputs them. It also doesn't attempt to parse values or do
any other interpretation, except for the equality checks (and it doesn't
do eq_1 or eq_2 since those would require parsing out the nybbles rather
than just splitting the hex string).

It also doesn't understand the `dbg!` construction although this looks
like it will not be too hard to add.

Should definitely extend and better-document the ExecTracer trait upstream.
But for now this is already pretty useful.