bump rust-simplicity to 0.7; update `simplicity pset run`
We retain the existing "only log jets in `simplicity pset run`" behavior,
although we now have the ability to log every node. My expectation is that
in practice users will usually only care about the flow of data between
jets. We can add extensions later to improve this.

Meanwhile, this greatly improves the output; rather than doing a hex dump
we output a full value decode which distinguishes left and right sums and
so on. We also split up the inputs to the eq_1 and eq_2 jets which we
previously didn't do because they'd be less than one hex nybble.